Renamed geocoder module
